fix some routes, refactor
