+++ src/components/httpRoutes/aquarius.ts
@@ -10,7 +10,9 @@ import { DDO } from '../../@types/DDO/DDO.js'
 
 export const aquariusRoutes = express.Router()
 
-aquariusRoutes.get('/assets/ddo/:did', async (req, res) => {
+export const AQUARIUS_API_BASE_PATH = '/api/aquarius'
+
+aquariusRoutes.get(`${AQUARIUS_API_BASE_PATH}/assets/ddo/:did`, async (req, res) => {
   try {
     const { did } = req.params
     if (!did || !did.startsWith('did:op')) {
@@ -34,7 +36,7 @@ aquariusRoutes.get('/assets/ddo/:did', async (req, res) => {
   }
 })
 
-aquariusRoutes.get('/assets/metadata/:did', async (req, res) => {
+aquariusRoutes.get(`${AQUARIUS_API_BASE_PATH}/assets/metadata/:did`, async (req, res) => {
   try {
     const { did } = req.params
     if (!did || !did.startsWith('did:op')) {
@@ -58,31 +60,34 @@ aquariusRoutes.get('/assets/metadata/:did', async (req, res) => {
   }
 })
 
-aquariusRoutes.post('/assets/metadata/query', async (req, res) => {
-  try {
-    const query = req.body
-    if (!query) {
-      res.status(400).send('Missing required body')
-      return
-    }
+aquariusRoutes.post(
+  `${AQUARIUS_API_BASE_PATH}/assets/metadata/query`,
+  async (req, res) => {
+    try {
+      const query = req.body
+      if (!query) {
+        res.status(400).send('Missing required body')
+        return
+      }
 
-    const result = await new QueryHandler(req.oceanNode).handle({
-      query,
-      command: PROTOCOL_COMMANDS.QUERY
-    })
-    if (result.stream) {
-      const queryResult = JSON.parse(await streamToString(result.stream as Readable))
-      res.json(queryResult)
-    } else {
-      res.status(result.status.httpStatus).send(result.status.error)
+      const result = await new QueryHandler(req.oceanNode).handle({
+        query,
+        command: PROTOCOL_COMMANDS.QUERY
+      })
+      if (result.stream) {
+        const queryResult = JSON.parse(await streamToString(result.stream as Readable))
+        res.json(queryResult)
+      } else {
+        res.status(result.status.httpStatus).send(result.status.error)
+      }
+    } catch (error) {
+      HTTP_LOGGER.log(LOG_LEVELS_STR.LEVEL_ERROR, `Error: ${error}`)
+      res.status(500).send('Internal Server Error')
     }
-  } catch (error) {
-    HTTP_LOGGER.log(LOG_LEVELS_STR.LEVEL_ERROR, `Error: ${error}`)
-    res.status(500).send('Internal Server Error')
   }
-})
+)
 
-aquariusRoutes.get('/state/ddo', async (req, res) => {
+aquariusRoutes.get(`${AQUARIUS_API_BASE_PATH}/state/ddo`, async (req, res) => {
   try {
     let query
     const did = String(req.query.did)
@@ -133,7 +138,7 @@ aquariusRoutes.get('/state/ddo', async (req, res) => {
   }
 })
 
-aquariusRoutes.post('/assets/ddo/validate', async (req, res) => {
+aquariusRoutes.post(`${AQUARIUS_API_BASE_PATH}/assets/ddo/validate`, async (req, res) => {
   try {
     if (!req.body || req.body === undefined) {
       res.status(400).send('Missing DDO object')
